+++ mobile/android/android-components/components/browser/toolbar/src/test/java/mozilla/components/browser/toolbar/BrowserToolbarTest.kt
@@ -4,13 +4,16 @@
 
 package mozilla.components.browser.toolbar
 
+import android.content.Context
 import android.graphics.Color
 import android.graphics.Typeface
 import android.graphics.drawable.Drawable
+import android.support.v13.view.inputmethod.EditorInfoCompat
 import android.util.AttributeSet
 import android.view.View
 import android.widget.ImageButton
 import android.widget.LinearLayout
+import androidx.test.core.app.ApplicationProvider
 import mozilla.components.browser.menu.BrowserMenuBuilder
 import mozilla.components.browser.toolbar.BrowserToolbar.Companion.ACTION_PADDING_DP
 import mozilla.components.browser.toolbar.display.DisplayToolbar
@@ -40,16 +43,19 @@ import org.robolectric.RuntimeEnvironment
 
 @RunWith(RobolectricTestRunner::class)
 class BrowserToolbarTest {
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
+    
     @Test
     fun `display toolbar is visible by default`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertTrue(toolbar.displayToolbar.visibility == View.VISIBLE)
         assertTrue(toolbar.editToolbar.visibility == View.GONE)
     }
 
     @Test
     fun `calling editMode() makes edit toolbar visible`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertTrue(toolbar.displayToolbar.visibility == View.VISIBLE)
         assertTrue(toolbar.editToolbar.visibility == View.GONE)
 
@@ -61,7 +67,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `calling displayMode() makes display toolbar visible`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.editMode()
 
         assertTrue(toolbar.displayToolbar.visibility == View.GONE)
@@ -75,7 +81,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `back presses will not be handled in display mode`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.displayMode()
 
         assertFalse(toolbar.onBackPressed())
@@ -86,7 +92,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `back presses will switch from edit mode to display mode`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.editMode()
 
         assertTrue(toolbar.displayToolbar.visibility == View.GONE)
@@ -100,7 +106,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `displayUrl will be forwarded to display toolbar immediately`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
         val ediToolbar = mock(EditToolbar::class.java)
 
@@ -115,7 +121,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `last URL will be forwarded to edit toolbar when switching mode`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val ediToolbar = mock(EditToolbar::class.java)
         toolbar.editToolbar = ediToolbar
@@ -130,7 +136,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `displayProgress will be forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
 
         toolbar.displayToolbar = displayToolbar
@@ -150,7 +156,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `internal onUrlEntered callback will be forwarded to urlChangeListener`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val mockedListener = object {
             var called = false
@@ -171,7 +177,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `toolbar measure will use full width and fixed 56dp height`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val widthSpec = View.MeasureSpec.makeMeasureSpec(1024, View.MeasureSpec.AT_MOST)
         val heightSpec = View.MeasureSpec.makeMeasureSpec(800, View.MeasureSpec.AT_MOST)
@@ -184,7 +190,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `toolbar will use provided height with EXACTLY measure spec`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val widthSpec = View.MeasureSpec.makeMeasureSpec(1024, View.MeasureSpec.AT_MOST)
         val heightSpec = View.MeasureSpec.makeMeasureSpec(800, View.MeasureSpec.EXACTLY)
@@ -197,7 +203,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `display and edit toolbar will use full size of browser toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertEquals(0, toolbar.displayToolbar.measuredWidth)
         assertEquals(0, toolbar.displayToolbar.measuredHeight)
@@ -217,7 +223,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `toolbar will switch back to display mode after an URL has been entered`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.editMode()
 
         assertTrue(toolbar.displayToolbar.visibility == View.GONE)
@@ -231,7 +237,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `display and edit toolbar will be laid out at the exact same position`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
         val editToolbar = mock(EditToolbar::class.java)
 
@@ -255,7 +261,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `menu builder will be forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertNull(toolbar.displayToolbar.menuBuilder)
 
@@ -268,7 +274,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `add browser action will be forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
 
         toolbar.displayToolbar = displayToolbar
@@ -284,7 +290,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `add page action will be forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val displayToolbar = mock(DisplayToolbar::class.java)
 
@@ -301,7 +307,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `URL update does not override search terms in edit mode`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
         val editToolbar = mock(EditToolbar::class.java)
 
@@ -323,7 +329,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `add navigation action will be forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
         toolbar.displayToolbar = displayToolbar
 
@@ -338,7 +344,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `invalidate actions is forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = mock(DisplayToolbar::class.java)
         toolbar.displayToolbar = displayToolbar
 
@@ -351,7 +357,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `search terms (if set) are forwarded to edit toolbar instead of URL`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val ediToolbar = mock(EditToolbar::class.java)
         toolbar.editToolbar = ediToolbar
@@ -370,7 +376,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `urlBoxBackgroundDrawable, browserActionMargin and urlBoxMargin are forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = toolbar.displayToolbar
 
         assertNull(displayToolbar.urlBoxView)
@@ -389,7 +395,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `onUrlClicked is forwarded to display toolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val displayToolbar = toolbar.displayToolbar
 
         assertTrue(displayToolbar.onUrlClicked())
@@ -401,15 +407,15 @@ class BrowserToolbarTest {
 
     @Test
     fun `layout of children will factor in padding`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.setPadding(50, 20, 60, 15)
         toolbar.removeAllViews()
 
-        val displayToolbar = spy(DisplayToolbar(RuntimeEnvironment.application, toolbar)).also {
+        val displayToolbar = spy(DisplayToolbar(context, toolbar)).also {
             toolbar.displayToolbar = it
         }
 
-        val editToolbar = spy(EditToolbar(RuntimeEnvironment.application, toolbar)).also {
+        val editToolbar = spy(EditToolbar(context, toolbar)).also {
             toolbar.editToolbar = it
         }
 
@@ -436,7 +442,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `editListener is set on EditToolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertNull(toolbar.editToolbar.editListener)
 
         val listener: Toolbar.OnEditListener = mock()
@@ -447,7 +453,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `editListener is invoked when switching between modes`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val listener: Toolbar.OnEditListener = mock()
         toolbar.setOnEditListener(listener)
@@ -465,7 +471,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `editListener is invoked when text changes`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         val listener: Toolbar.OnEditListener = mock()
         toolbar.setOnEditListener(listener)
@@ -485,7 +491,7 @@ class BrowserToolbarTest {
     @Test
     fun `BrowserToolbar Button must set padding`() {
         var button = BrowserToolbar.Button(mock(), "imageResource", visible = { true }) {}
-        val linearLayout = LinearLayout(RuntimeEnvironment.application)
+        val linearLayout = LinearLayout(context)
         var view = button.createView(linearLayout)
         val padding = Padding(0, 0, 0, 0)
         assertEquals(view.paddingLeft, ACTION_PADDING_DP)
@@ -527,7 +533,7 @@ class BrowserToolbarTest {
             selected = false,
             background = 0
         ) {}
-        val linearLayout = LinearLayout(RuntimeEnvironment.application)
+        val linearLayout = LinearLayout(context)
         var view = button.createView(linearLayout)
         assertEquals(view.paddingLeft, ACTION_PADDING_DP)
         assertEquals(view.paddingTop, ACTION_PADDING_DP)
@@ -565,7 +571,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `hint changes edit and display urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertNull(toolbar.displayToolbar.urlView.hint)
         assertNull(toolbar.editToolbar.urlView.hint)
@@ -580,7 +586,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `hintColor changes edit and display urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertTrue(toolbar.displayToolbar.urlView.currentHintTextColor != Color.RED)
         assertTrue(toolbar.editToolbar.urlView.currentHintTextColor != Color.RED)
@@ -593,7 +599,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `textColor changes edit and display urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertTrue(toolbar.displayToolbar.urlView.currentTextColor != Color.RED)
         assertTrue(toolbar.editToolbar.urlView.currentTextColor != Color.RED)
@@ -606,7 +612,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `textSize changes edit and display urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
 
         assertTrue(toolbar.displayToolbar.urlView.textSize != 12f)
         assertTrue(toolbar.editToolbar.urlView.textSize != 12f)
@@ -619,7 +625,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `typeface changes edit and display urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         val typeface: Typeface = mock()
 
         assertNotEquals(typeface, toolbar.editToolbar.urlView.typeface)
@@ -635,7 +641,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `obtainAttributes called when attributes provided`() {
-        val application = spy(RuntimeEnvironment.application)
+        val application = spy(context)
         val attributeSet: AttributeSet = mock()
 
         BrowserToolbar(application)
@@ -647,7 +653,7 @@ class BrowserToolbarTest {
 
     @Test
     fun `displaySiteSecurityIcon getter and setter`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertEquals(toolbar.displayToolbar.siteSecurityIconView.isVisible(), toolbar.displaySiteSecurityIcon)
 
         toolbar.displaySiteSecurityIcon = false
@@ -659,31 +665,31 @@ class BrowserToolbarTest {
 
     @Test
     fun `urlBoxView getter`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertEquals(toolbar.displayToolbar.urlBoxView, toolbar.urlBoxView)
     }
 
     @Test
     fun `browserActionMargin getter`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertEquals(toolbar.displayToolbar.browserActionMargin, toolbar.browserActionMargin)
     }
 
     @Test
     fun `urlBoxMargin getter`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertEquals(toolbar.displayToolbar.urlBoxMargin, toolbar.urlBoxMargin)
     }
 
     @Test
     fun `onUrlClicked getter`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         assertEquals(toolbar.displayToolbar.onUrlClicked, toolbar.onUrlClicked)
     }
 
     @Test
     fun `setUrlTextPadding applies padding to urlView`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.setUrlTextPadding(5, 5, 5, 5)
         assertEquals(5, toolbar.displayToolbar.urlView.paddingLeft)
         assertEquals(5, toolbar.displayToolbar.urlView.paddingTop)
@@ -741,12 +747,35 @@ class BrowserToolbarTest {
 
     @Test
     fun `siteSecure updates the displayToolbar`() {
-        val toolbar = BrowserToolbar(RuntimeEnvironment.application)
+        val toolbar = BrowserToolbar(context)
         toolbar.displayToolbar = spy(toolbar.displayToolbar)
         assertEquals(SiteSecurity.INSECURE, toolbar.siteSecure)
 
         toolbar.siteSecure = SiteSecurity.SECURE
 
         verify(toolbar.displayToolbar).setSiteSecurity(SiteSecurity.SECURE)
+    }
+
+    @Test
+    fun `private flag sets IME_FLAG_NO_PERSONALIZED_LEARNING on url edit view`() {
+        val toolbar = BrowserToolbar(context)
+        val editToolbar = toolbar.editToolbar
+
+        // By default "private mode" is off.
+        assertEquals(0, editToolbar.urlView.imeOptions and
+            EditorInfoCompat.IME_FLAG_NO_PERSONALIZED_LEARNING)
+        assertEquals(false, toolbar.private)
+
+        // Turning on private mode sets flag
+        toolbar.private = true
+        assertNotEquals(0, editToolbar.urlView.imeOptions and
+            EditorInfoCompat.IME_FLAG_NO_PERSONALIZED_LEARNING)
+        assertTrue(toolbar.private)
+
+        // Turning private mode off again - should remove flag
+        toolbar.private = false
+        assertEquals(0, editToolbar.urlView.imeOptions and
+            EditorInfoCompat.IME_FLAG_NO_PERSONALIZED_LEARNING)
+        assertEquals(false, toolbar.private)
     }
 }